renamed service class, and reorganized packages
