+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/MovieRepository.kt
@@ -1,10 +1,10 @@
 package com.mayarafernandes.movieplayer.movieList
 
-import com.mayarafernandes.movieplayer.movieList.service.TempService
+import com.mayarafernandes.movieplayer.movieList.service.MovieServiceIMPL
 import com.mayarafernandes.movieplayer.movieList.storage.LocalMovieStorage
 
 class MovieRepository (private val localMovieStorage: LocalMovieStorage) {
-    private val tempService = TempService()
+    private val tempService = MovieServiceIMPL()
 
     fun returnMovieList(): List<Movie> {
         val movieList = tempService.returnMovieList()
