+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/service/MovieServiceIMPL.kt
@@ -2,7 +2,7 @@ package com.mayarafernandes.movieplayer.movieList.service
 
 import com.mayarafernandes.movieplayer.movieList.Supplier
 
-class TempService : MovieService {
+class MovieServiceIMPL : MovieService {
     private val movieList = Supplier.movies
 
     override fun returnMovieList() = movieList
