+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/view/MovieListActivity.kt
@@ -9,7 +9,8 @@ import com.mayarafernandes.movieplayer.movieList.MovieRepository
 import com.mayarafernandes.movieplayer.movieList.storage.MemoryRepository
 import kotlinx.android.synthetic.main.activity_movie_list.*
 
-class MovieListActivity : AppCompatActivity(), MovieListView, MovieViewModelClickListener {
+class MovieListActivity : AppCompatActivity(),
+    MovieListView, MovieViewModelClickListener {
 
     private lateinit var controller: MovieListController
