+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/view/MovieListView.kt
@@ -1,4 +1,4 @@
-package com.mayarafernandes.movieplayer.movieList
+package com.mayarafernandes.movieplayer.movieList.view
 
 import com.mayarafernandes.movieplayer.movieList.view.MovieViewModel
